+++ server/src/test/java/org/opensearch/indices/cluster/IndicesClusterStateServiceRandomUpdatesTests.java
@@ -584,6 +584,7 @@ public class IndicesClusterStateServiceRandomUpdatesTests extends AbstractIndice
             primaryReplicaSyncer,
             s -> {},
             RetentionLeaseSyncer.EMPTY,
+            null,
             null
         );
     }